feat(kubelet): Fully glues together logs logic
As part of this, I changed the provider trait to instead return a vec
of bytes instead of worrying about all the allocations of copying each
line into a string. Instead the raw data is just passed back as part of
the HTTP response body.

I also found that the wasi logs implementation needed to reset the seek
pointer to the beginning of the file for future reads, so that has now
been fixed

NOTE: This is not currently tested with a running master because we do
not have HTTPS set up yet. It has been tested with manual calls to the
endpoint

Closes #33