fix(justfile): Fixes "wrong password" issue with generated certs
Based on testing from the team and what I found on the internet (always
a reliable source of information), having an empty password with a
pkcs bundle can cause errors. This exports a default password to be
used with the certificate.

Also, I removed the kubelet=debug setting for logging because kubelet is
a _really_ verbose default