downgrade to azure-cli 2.9.1
A recent bug was introduced to the azure CLI preventing `az storage blob upload-batch` from working. In the meantime, a hotfix has been provided to the HEAD of azure-blob-storage-upload reverting back to azure-cli v2.9.1

The `sync` flag was also introduced since v1.0.0, which is enabled by default. sync must be disabled to preserve files currently present in the storage account.

Signed-off-by: Matthew Fisher <matt.fisher@microsoft.com>