ref(volume): Streamlines how volumes are mounted and unmounted
This builds on #584 (well actually tears down the original idea, but
you get it) and makes volume management all handled by the same object.
This is a breaking change that removes the `Ref` type and replaces it
with a fully managed `VolumeRef` type. I am not entirely happy with having
the unmounting being handled in the async_drop stage, but it is good enough
for now. It wouldn't be a breaking change to make it a new state down the line
if we want to move it.

Closes #585