refactor: get rid of stale dependencies
Remove the following dependencies:
  * hyperx
  * www-authenticate

The `hyperx` crate doesn't seem to be maintained anymore. Its
dependencies have upper bounds that are causing upgrade conflicts both
to oci-distribution and to all the consumers of it.
See https://github.com/dekellum/hyperx/issues/39

hyperx is a fork of a really old version of the popular hyper crate. We
require it just to be able to consume the `www-authenticate` crate,
which depends on types that are defined inside of hyperx. These types
have been removed from the official hyper crate.

This commit replaces these two dependencies with a more modern one
called `http-auth`, which builds on top of the popular `http` crate.

This commit also extends the test suite to perform checks against a docker
registry that uses basic authentication.

No regression have been introduced by the refeactoring.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>