fix: address broken test
Fix one of our unit test that suddenly started to fail.

The test was doing to following operations:

1. Start local OCI registry configured to require authentication
2. Pull a wasm module from the public Azure Container Registry
3. Push the wasm module into the local OCI registry

The test was using the local OCI registry credentials to authenticate
against the Azure Container Registry. This was not needed because the module
being pulled is public, hence no auth is required.
Something changed with regards to Azure CR, which started to reject our
authentication efforts.

The test has been changed to perform an anonymous pull from Azure CR.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>