ref(client)!: create new LayerDescriptor for pulling blobs
The old OciDescriptor has too many fields, which maybe aren't available
when trying to pull a blob with only the digest. In the past you only
needed the digest as string, but the optional urls field was added. But
all other fields of the OciDescriptor aren't required when pulling a
blob, but are required by the struct. The LayerDescriptor struct now has
only the fields which are used for pulling.

This is a breaking change, as it changes the signature of the pull
methods.

Signed-off-by: Benjamin Neff <benjamin@coding4coffee.ch>