fix(client): Validate digests when header is set
This implements missing functionality around using the `Docker-Content-Digest`
header to validate the digest of incoming content. This should now be
spec compliant.

Additionally, as a safety measure, if a digest is given as part of a
image reference, that digest will also be used if the header isn't set
or it will be validated if the algorithm is the same

Fixes #149

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>