Filter empty DOCKER_DIGEST_HEADER.
The spec says the header MUST be used if "if present on the response",
but does not define what present means. If present means "the key is
there", then the current behavior is correct and the server is not spec
compliant. However, if it means "the key is there and not empty", then
the client needs to be more forgiving than it currently is.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>